pyproject.toml
[project]
name = "whisper-pasts"
version = "2.0.0"
description = "语音转文字工具: 通过快捷键录音并转录为文字"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "whisper-pasts team" }
]
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "scipy>=1.10.0",
    "pynput>=1.7.6",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
whisper-pasts = "main:main"

[tool.uv]
sources = {}